requirements/tests.txt
flake8
coverage
django-reverson